+++ internal/cmd/self.go
@@ -26,5 +26,6 @@ func initSelfCmd(rootCmd *cobra.Command, version string, openURL func(string) er
 
 	selfCmd.AddCommand(initDocCommand(openURL))
 	selfCmd.AddCommand(initLspCommand(version))
+	selfCmd.AddCommand(initSkillsCommand())
 	selfCmd.AddCommand(initUpgradeCommand(version))
 }